# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "GPS Tracking History",
      "description": "GPS Tracking History"
     
    },
    "host": "localhost:3000",
    "basePath": "",
...
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/getUserData/{userId}": {
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "ID of User that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Get User with given ID",
          "responses": {
            "200": {
              "description": "User is found",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "/getUserTracks/{userId}": {
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "ID of User to find his tracks",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Users"
          ],
          "summary": "Get Tracks with given user ID",
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Tracks"
              }
            }
          }
        }
      },
...
  },
    "definitions": {
      "Track": {
        "required": [
        
          "_id"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "format":"ObjectId",
            "uniqueItems": true
          },
          "trackId": {
            "type": "string",
            "uniqueItems": true
          },
          "startTime": {
            "type": "string",
            "format":"date-time"
          },
          "endTime": {
            "type": "string",
            "format":"date-time"
          },
          "userID":{
            "type": "string",
            "format":"ObjectId"
          }
        }
      },
...
      "Tracks": {
        "type": "array",
        "$ref": "#/definitions/Track"
      },
      "User": {
        "required": [
        
          "_id"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "format":"ObjectId",
            "uniqueItems": true
          },
          "userName": {
            "type": "string"
          }
        }
      },
...
    }
  }